// src/app/generated_services/model/carlonGracieBackendProgressionApplicationDTOsShowGraduationDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CarlonGracieBackendSharedDomainEnumsStripesEnum } from './carlonGracieBackendSharedDomainEnumsStripesEnum';
import { CarlonGracieBackendProgressionApplicationDTOsShowBeltDTO } from './carlonGracieBackendProgressionApplicationDTOsShowBeltDTO';


export interface CarlonGracieBackendProgressionApplicationDTOsShowGraduationDTO { 
    id?: string;
    createdAt?: string;
    belt?: CarlonGracieBackendProgressionApplicationDTOsShowBeltDTO | null;
    studentUserName?: string | null;
    studentFullName?: string | null;
    studentId: string;
    beltId: string;
    graduationDate: string;
    stripes?: CarlonGracieBackendSharedDomainEnumsStripesEnum;
}
export namespace CarlonGracieBackendProgressionApplicationDTOsShowGraduationDTO {
}