// src/app/generated_services/model/carlonGracieBackendProgressionApplicationDTOsUpdateGraduationDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CarlonGracieBackendSharedDomainEnumsStripesEnum } from './carlonGracieBackendSharedDomainEnumsStripesEnum';


export interface CarlonGracieBackendProgressionApplicationDTOsUpdateGraduationDTO { 
    studentId: string;
    beltId: string;
    graduationDate: string;
    stripes?: CarlonGracieBackendSharedDomainEnumsStripesEnum;
}
export namespace CarlonGracieBackendProgressionApplicationDTOsUpdateGraduationDTO {
}